gm-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "gm-server"
version = "0.1.0"
description = "LLM-based Game Master server for Arma 3"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gm_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/gm_server/tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
